test(emit): split class member tests
